# docs/api/User_Api_Swagger.yaml
...
openapi: 3.0.0
info:
  title: ESP RainMaker Neo Auth & Accounts API
  version: 0.8.0
  description: |
    Sign-in and account management for the ESP RainMaker Neo cloud.
    Anything that produces or rotates an access token lives here; once
    you have a token, the rest of the platform lives in the
    **ESP RainMaker Neo Platform API**

    ESP User is the platform's own OAuth 2.1 / OIDC issuer: every token a client
    holds is minted here, regardless of how the user proved their identity.
    Admins authenticate directly against their AWS Cognito user pool.

    Login methods (all end in OUR tokens; upstream credentials never reach clients):

    - **Federated login (default)**: `GET /oauth2/authorize` starts a browser
      authorization-code + PKCE flow. With a single enabled identity provider the
      user is auto-redirected to it; the
      broker exchanges the upstream code server-side and issues OUR authorization
      code at the client's `redirect_uri`.
    - **Native password APIs** (`/v1/user/auth/*`): the main-era
      native surface. Credentials are verified against the bundled Cognito pool server-side and converted to our tokens.

    Token lifecycle after login:

    1. `POST /oauth2/token` (`grant_type=refresh_token`) - rotate the token set
       (native `/v1/user/auth/token/refresh` is reuse-tolerant instead).
    2. Send `access_token` as a `Bearer` `Authorization` header on subsequent
...
       must come from the same sign-in.
...
  contact:
...
  license:
...

...
servers:
...
paths:
...
  /oauth2/federation/start:
    get:
      summary: Redirect to an upstream identity provider
      description: |
        Sends the browser to the upstream provider's login. Reads the flow from the `__Host-esp_flow_id` cookie, refuses a provider the client does not allow, and records this leg's own `state`, `nonce` and PKCE verifier on the flow. The client's values never reach the provider.
      tags:
        - Federated Login
      security: []
      parameters:
        - name: provider
          in: query
          required: true
          description:
            Provider name from the identity-provider registry (e.g. `cognito`).
          schema:
            type: string
        - name: __Host-esp_flow_id
          in: cookie
          required: true
          description: Login-flow session cookie set by `/oauth2/authorize`.
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the provider's hosted login page.
          headers:
            Location:
              schema:
                type: string
        "400":
          description:
            Missing/expired flow or unknown provider (HTML error page).
          content:
            text/html:
              schema:
                type: string
  /oauth2/federation/callback:
    get:
      summary: Upstream provider callback
      description: |
        The provider returns here after login. Verifies `state` against the flow, exchanges the upstream code, validates the upstream `id_token`, finds or creates the ESP User account, and redirects to the client's `redirect_uri` with a single-use `code`. Upstream tokens are discarded.
      tags:
        - Federated Login
      security: []
      parameters:
        - name: code
          in: query
          required: true
          description: Upstream authorization code.
          schema:
            type: string
        - name: state
          in: query
          required: true
          description: The broker's own HMAC-bound state (not the client's).
          schema:
            type: string
      responses:
        "302":
          description:
            Redirect to the client's `redirect_uri` with `code` (+ `state`).
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              description: |
                The `__Host-esp_session` cookie
                (`Path=/; HttpOnly; Secure; SameSite=Lax`) establishing the
                authorization server's own session, whose record inherits the
                upstream's `auth_time`. It stores no token — an opaque revocable
                identifier only. A session-layer failure never blocks the login
                itself.
              schema:
                type: string
        "400":
          description:
            State/flow mismatch or upstream validation failure (HTML error
            page).
          content:
            text/html:
              schema:
                type: string
...
components:
...
  securitySchemes:
    CognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      description: |
        Admin Cognito User Pool JWT, for the admin APIs (`/v1/admin/*`).
        End-user routes take an ESP User token instead — see `bearerAuth`.
      x-amazon-apigateway-authtype: cognito_user_pools
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An ESP User access token, sent as a Bearer token. The token is verified
        against the provider's published public keys.
    clientBasicAuth:
      type: http
      scheme: basic
      description: |
        OAuth client authentication (RFC 6749 §2.3.1) as HTTP Basic
        `base64(client_id:client_secret)`. A confidential client sends its secret;
        a public client sends its client_id with an empty secret.
security: []